Reworked how the content and inventories table were stored by Stockpile. It now uses an array of floppy disks to read and write large file to and from which unlock virtually an unlimited storage capacity (up to tens of millions of items possibly). It works by splitting tables into multiple chunks and save each chunk to a floppy disk.
I updated the doc to explain the change and provide a quick guide on how to set up floppy disks.

Change to the Unit system, now when running the list_all_inventories() API method, it will add inventories which are not part of user defined units to the default "unassigned" unit.
This allows for easier unit management for clients.